mod _rlp extracted
