+++ rust/signer/src/_rlp.rs
@@ -1,0 +1,24 @@
+use rustc_serialize::hex::{ToHex, FromHex};
+use string::StringPtr;
+use rlp::decode_list;
+
+// RLP
+
+fn safe_rlp_item(rlp: &str, position: u32) -> Result<String, String> {
+  let hex = rlp.from_hex().map_err(| e | e.to_string())?;
+  let rlp = decode_list::<Vec<u8>>(&hex);
+  let data = rlp.get(position as usize).ok_or_else(|| "Invalid RLP position".to_string())?;
+  Ok(data.to_hex())
+}
+
+#[no_mangle]
+pub unsafe extern fn rlp_item(rlp: *mut StringPtr, position: u32, error: *mut u32) -> *mut String {
+  match safe_rlp_item((*rlp).as_str(), position) {
+    Ok(result) => Box::into_raw(Box::new(result)),
+    Err(_err) => {
+      *error = 1;
+      let s: String = "".into();
+      Box::into_raw(Box::new(s))
+    }
+  }
+}
