+++ rust/signer/src/eth.rs
@@ -2,7 +2,6 @@ use wordlist;
 use rustc_serialize::hex::{ToHex, FromHex};
 use rustc_serialize::base64::{self, ToBase64};
 use blockies::{Blockies, create_icon, ethereum};
-use rlp::decode_list;
 use string::StringPtr;
 use tiny_keccak::Keccak;
 use ethkey::{KeyPair, Generator, Brain, Message, sign};
@@ -64,26 +63,7 @@ pub unsafe extern fn ethkey_keypair_sign(keypair: *mut KeyPair, message: *mut St
   Box::into_raw(Box::new(signature))
 }
 
-// RLP
 
-fn safe_rlp_item(rlp: &str, position: u32) -> Result<String, String> {
-  let hex = rlp.from_hex().map_err(| e | e.to_string())?;
-  let rlp = decode_list::<Vec<u8>>(&hex);
-  let data = rlp.get(position as usize).ok_or_else(|| "Invalid RLP position".to_string())?;
-  Ok(data.to_hex())
-}
-
-#[no_mangle]
-pub unsafe extern fn rlp_item(rlp: *mut StringPtr, position: u32, error: *mut u32) -> *mut String {
-  match safe_rlp_item((*rlp).as_str(), position) {
-    Ok(result) => Box::into_raw(Box::new(result)),
-    Err(_err) => {
-      *error = 1;
-      let s: String = "".into();
-      Box::into_raw(Box::new(s))
-    }
-  }
-}
 
 fn blockies_icon_in_base64(seed: Vec<u8>) -> String {
   let mut result = Vec::new();
